test(http): share recorded response registrations
