+++ tests/test_patronus_evaluate_api.py
@@ -13,38 +13,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import json
-
 import pytest
 
 from nemoguardrails import RailsConfig
 from nemoguardrails.actions.actions import ActionResult, action
-from nemoguardrails.http import HTTPResponse
-from nemoguardrails.http.testing import RecordingHTTPClient
+from nemoguardrails.http.testing import RecordedHTTPResponses
 from nemoguardrails.library.patronusai.actions import (
     check_guardrail_pass,
     patronus_evaluate_request,
 )
 from tests.utils import TestChat
-
-
-class RecordedHTTPResponses:
-    def __init__(self):
-        self.client = RecordingHTTPClient()
-        self.urls = []
-
-    def post(self, url, *, payload=None, status=200, body=None):
-        content = body.encode() if body is not None else json.dumps(payload).encode()
-        self.client.add_response(HTTPResponse(status_code=status, content=content))
-        self.urls.append(url)
-
-    def __enter__(self):
-        return self
-
-    def __exit__(self, exc_type, exc_value, traceback):
-        if exc_type is None:
-            assert all(request.url in self.urls for request in self.client.requests)
-
 
 PATRONUS_EVALUATE_API_URL = "https://api.patronus.ai/v1/evaluate"
 COLANG_CONFIG = """
@@ -380,32 +358,10 @@ def test_patronus_evaluate_api_internal_error_when_no_env_set():
     with RecordedHTTPResponses() as m:
         chat.app.register_action_param("http_client", m.client)
         chat.app.register_action(retrieve_relevant_chunks, "retrieve_relevant_chunks")
-        m.post(
-            PATRONUS_EVALUATE_API_URL,
-            payload={
-                "results": [
-                    {
-                        "evaluator_id": "lynx-large-2024-07-23",
-                        "criteria": "patronus:hallucination",
-                        "status": "success",
-                        "evaluation_result": {
-                            "pass": False,
-                        },
-                    },
-                    {
-                        "evaluator_id": "answer-relevance-large-2024-07-23",
-                        "criteria": "patronus:answer-relevance",
-                        "status": "success",
-                        "evaluation_result": {
-                            "pass": False,
-                        },
-                    },
-                ]
-            },
-        )
 
         chat >> "Hi"
         chat << "I'm sorry, an internal error has occurred."
+        assert m.client.requests == []
 
 
 def test_patronus_evaluate_api_internal_error_when_no_evaluators_provided():
@@ -437,32 +393,10 @@ def test_patronus_evaluate_api_internal_error_when_no_evaluators_provided():
     with RecordedHTTPResponses() as m:
         chat.app.register_action_param("http_client", m.client)
         chat.app.register_action(retrieve_relevant_chunks, "retrieve_relevant_chunks")
-        m.post(
-            PATRONUS_EVALUATE_API_URL,
-            payload={
-                "results": [
-                    {
-                        "evaluator_id": "lynx-large-2024-07-23",
-                        "criteria": "patronus:hallucination",
-                        "status": "success",
-                        "evaluation_result": {
-                            "pass": False,
-                        },
-                    },
-                    {
-                        "evaluator_id": "answer-relevance-large-2024-07-23",
-                        "criteria": "patronus:answer-relevance",
-                        "status": "success",
-                        "evaluation_result": {
-                            "pass": False,
-                        },
-                    },
-                ]
-            },
-        )
 
         chat >> "Hi"
         chat << "I'm sorry, an internal error has occurred."
+        assert m.client.requests == []
 
 
 def test_patronus_evaluate_api_internal_error_when_evaluator_dict_does_not_have_evaluator_key():
@@ -501,32 +435,10 @@ def test_patronus_evaluate_api_internal_error_when_evaluator_dict_does_not_have_
     with RecordedHTTPResponses() as m:
         chat.app.register_action_param("http_client", m.client)
         chat.app.register_action(retrieve_relevant_chunks, "retrieve_relevant_chunks")
-        m.post(
-            PATRONUS_EVALUATE_API_URL,
-            payload={
-                "results": [
-                    {
-                        "evaluator_id": "lynx-large-2024-07-23",
-                        "criteria": "patronus:hallucination",
-                        "status": "success",
-                        "evaluation_result": {
-                            "pass": False,
-                        },
-                    },
-                    {
-                        "evaluator_id": "answer-relevance-large-2024-07-23",
-                        "criteria": "patronus:answer-relevance",
-                        "status": "success",
-                        "evaluation_result": {
-                            "pass": False,
-                        },
-                    },
-                ]
-            },
-        )
 
         chat >> "Hi"
         chat << "I'm sorry, an internal error has occurred."
+        assert m.client.requests == []
 
 
 @pytest.mark.asyncio
